% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper.R
\docType{data}
\name{riverforest}
\alias{riverforest}
\title{River Forest, IL, Property Tax Data}
\format{An data.frame (or data.table) object}
\usage{
data(riverforest)
}
\description{
Data has been scraped downloaded from the Cook County, IL, property tax website
at <https://www.cookcountyassessor.com/Search/Property-Owner-Search.aspx> using
combination of a) Property Index Number (PIN) search via the URL
\url{http://www.cookcountyassessor.com/Search/Property-Search.aspx} and
b) per-prortery access given a pin (as final argument) via
\url{http://www.cookcountyassessor.com/Property.aspx?mode=details&pin=...}.
The resulting data set contains all columns provided by the website.
}
\keyword{datasets}
